refactor v1 expression when helpers
